requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
lxml==6.0.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
regex==2025.7.34
requests==2.32.4
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
tiktoken==0.10.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
yt-dlp==2025.7.21
